Refactor Create Project page into separate Form Workflow component
